--- CKDX9Rasterizer/CKDX9PixelShaderDesc.cpp
@@ -1,10 +1,0 @@
-#include "CKDX9Rasterizer.h"
-
-CKBOOL CKDX9PixelShaderDesc::Create(CKDX9RasterizerContext *Ctx, CKDWORD *Function)
-{
-    Owner = Ctx;
-    m_Function = Function;
-    return SUCCEEDED(Ctx->m_Device->SetPixelShader(DxShader));
-}
-
-CKDX9PixelShaderDesc::~CKDX9PixelShaderDesc() = default;
